+++ chainer/functions/__init__.py
@@ -16,13 +16,9 @@ from chainer.functions.array import split_axis
 from chainer.functions.connection import bilinear
 from chainer.functions.connection import convolution_2d
 from chainer.functions.connection import embed_id
-from chainer.functions.connection import inception
-from chainer.functions.connection import inceptionbn
 from chainer.functions.connection import linear
-from chainer.functions.connection import parameter
 from chainer.functions.evaluation import accuracy
 from chainer.functions.loss import cross_covariance
-from chainer.functions.loss import hierarchical_softmax
 from chainer.functions.loss import mean_squared_error
 from chainer.functions.loss import negative_sampling
 from chainer.functions.loss import sigmoid_cross_entropy
@@ -41,6 +37,17 @@ from chainer.functions.normalization import local_response_normalization
 from chainer.functions.pooling import average_pooling_2d
 from chainer.functions.pooling import max_pooling_2d
 from chainer.functions.pooling import spatial_pyramid_pooling_2d
+import chainer.links.connection.bilinear as links_bilinear
+import chainer.links.connection.convolution_2d as links_convolution_2d
+import chainer.links.connection.embed_id as links_embed_id
+from chainer.links.connection import inception
+from chainer.links.connection import inceptionbn
+import chainer.links.connection.linear as links_linear
+import chainer.links.normalization.batch_normalization \
+    as links_batch_normalization
+from chainer.links.connection import parameter
+from chainer.links.loss import hierarchical_softmax
+import chainer.links.loss.negative_sampling as links_negative_sampling
 
 
 ClippedReLU = clipped_relu.ClippedReLU
@@ -70,28 +77,21 @@ reshape = reshape.reshape
 SplitAxis = split_axis.SplitAxis
 split_axis = split_axis.split_axis
 
-Bilinear = bilinear.Bilinear
-Convolution2D = convolution_2d.Convolution2D
 convolution_2d = convolution_2d.convolution_2d
-EmbedID = embed_id.EmbedID
-Inception = inception.Inception
-InceptionBN = inceptionbn.InceptionBN
-Linear = linear.Linear
+embed_id = embed_id.embed_id
 linear = linear.linear
-Parameter = parameter.Parameter
 
 Accuracy = accuracy.Accuracy
 accuracy = accuracy.accuracy
 
 bernoulli_nll = vae.bernoulli_nll
-BinaryHierarchicalSoftmax = hierarchical_softmax.BinaryHierarchicalSoftmax
 CrossCovariance = cross_covariance.CrossCovariance
 cross_covariance = cross_covariance.cross_covariance
 gaussian_kl_divergence = vae.gaussian_kl_divergence
 gaussian_nll = vae.gaussian_nll
 MeanSquaredError = mean_squared_error.MeanSquaredError
 mean_squared_error = mean_squared_error.mean_squared_error
-NegativeSampling = negative_sampling.NegativeSampling
+negative_sampling = negative_sampling.negative_sampling
 SigmoidCrossEntropy = sigmoid_cross_entropy.SigmoidCrossEntropy
 sigmoid_cross_entropy = sigmoid_cross_entropy.sigmoid_cross_entropy
 SoftmaxCrossEntropy = softmax_cross_entropy.SoftmaxCrossEntropy
@@ -119,7 +119,8 @@ dropout = dropout.dropout
 Gaussian = gaussian.Gaussian
 gaussian = gaussian.gaussian
 
-BatchNormalization = batch_normalization.BatchNormalization
+fixed_batch_normalization = batch_normalization.fixed_batch_normalization
+batch_normalization = batch_normalization.batch_normalization
 LocalResponseNormalization = \
     local_response_normalization.LocalResponseNormalization
 local_response_normalization = \
@@ -132,3 +133,19 @@ max_pooling_2d = max_pooling_2d.max_pooling_2d
 SpatialPyramidPooling2D = spatial_pyramid_pooling_2d.SpatialPyramidPooling2D
 spatial_pyramid_pooling_2d = \
     spatial_pyramid_pooling_2d.spatial_pyramid_pooling_2d
+
+
+# Left for backward compatibility
+
+Bilinear = links_bilinear.Bilinear
+Convolution2D = links_convolution_2d.Convolution2D
+EmbedID = links_embed_id.EmbedID
+Inception = inception.Inception
+InceptionBN = inceptionbn.InceptionBN
+Linear = links_linear.Linear
+Parameter = parameter.Parameter
+
+BatchNormalization = links_batch_normalization.BatchNormalization
+
+BinaryHierarchicalSoftmax = hierarchical_softmax.BinaryHierarchicalSoftmax
+NegativeSampling = links_negative_sampling.NegativeSampling
