+++ chainer/functions/connection/bilinear.py
@@ -117,8 +117,41 @@ class BilinearFunction(function.Function):
 def bilinear(e1, e2, W, V1=None, V2=None, b=None):
     """Applies a bilinear function based on given parameters.
 
-    Note that V1, V2, b are optional, though they must not be None if at least
-    one of them is not None.
+    This is a building block of Neural Tensor Network (see the reference paper
+    below). It takes two input variables and one or four parameters, and
+    outputs one variable.
+
+    To be precise, denote six input arrays mathematically by
+    :math:`e^1\in \mathbb{R}^{I\cdot J}`,
+    :math:`e^2\in \mathbb{R}^{I\cdot K}`,
+    :math:`W\in \mathbb{R}^{J \cdot K \cdot L}`,
+    :math:`V^1\in \mathbb{R}^{J \cdot L}`,
+    :math:`V^2\in \mathbb{R}^{K \cdot L}`, and :math:`b\in \mathbb{R}^{L}`.
+    where :math:`I` is mini-batch size.
+    In this document, we call :math:`V^1`, :math:`V^2`, and :math:`b` linear
+    parameters.
+
+    The output of forward propagation is calculated as
+
+    .. math::
+
+      y_{il} = \sum_{jk} e^1_{ij} e^2_{ik} W_{jkl} + \
+        \sum_{j} e^1_{ij} V^1_{jl} + \sum_{k} e^2_{ik} V^2_{kl} + b_{l}.
+
+    Note that V1, V2, b are optional. If these are not given, then this
+    function omits the last three terms in the above equation.
+
+    .. note::
+
+       This function accepts an input variable ``e1`` or ``e2`` of a non-matrix
+       array. In this case, the leading dimension is treated as the batch
+       dimension, and the other dimensions are reduced to one dimension.
+
+    .. note::
+
+       In the original paper, :math:`J` and :math:`K`
+       must be equal and the author denotes :math:`[V^1 V^2]`
+       (concatenation of matrices) by :math:`V`.
 
     Args:
         e1 (~chainer.Variable): Left input variable.
@@ -128,8 +161,13 @@ def bilinear(e1, e2, W, V1=None, V2=None, b=None):
         V2 (~chainer.Variable): Right coefficient variable.
         b (~chainer.Variable): Bias variable.
 
-    Reutnrs:
+    Returns:
         ~chainer.Variable: Output variable.
+
+    See:
+        `Reasoning With Neural Tensor Networks for Knowledge Base Completion
+        <http://papers.nips.cc/paper/5028-reasoning-with-neural-tensor-
+        networks-for-knowledge-base-completion>`_ [Socher+, NIPS2013].
 
     """
     flags = [V1 is None, V2 is None, b is None]
@@ -140,117 +178,3 @@ def bilinear(e1, e2, W, V1=None, V2=None, b=None):
         return BilinearFunction()(e1, e2, W)
     else:
         return BilinearFunction()(e1, e2, W, V1, V2, b)
-
-
-class Bilinear(link.Link):
-
-    """Bilinear function, an extension of Linear function.
-
-    ``Bilinear`` function takes two input vectors and outputs one vector.
-    If one of the input vectors is fixed, this function works
-    as an affine transform of the other input vector.
-
-    ``Bilinear`` function is a building block of Neural Tensor Network
-    (See the reference paper below).
-
-    To be precise, ``Bilinear`` function has four parameters,
-    :math:`W\in \mathbb{R}^{J \cdot K \cdot L}`,
-    :math:`V^1\in \mathbb{R}^{J \cdot L}`,
-    :math:`V^2\in \mathbb{R}^{K \cdot L}`, and :math:`b\in \mathbb{R}^{L}`.
-    In this document, we call :math:`V^1`, :math:`V^2`,
-    and :math:`b` linear parameters.
-
-    Given two inputs (in a mini-batch manner)
-    :math:`e^1\in \mathbb{R}^{I\cdot J}` and
-    :math:`e^2\in \mathbb{R}^{I\cdot K}`
-    where :math:`I` is mini-batch size, the output of forward propagation is
-    calculated as
-
-    .. math::
-
-      y_{il} = \sum_{jk} e^1_{ij} e^2_{ik} W_{jkl} + \
-        \sum_{j} e^1_{ij} V^1_{jl} + \sum_{k} e^2_{ik} V^2_{kl} + b_{l}.
-
-    If ``nobias`` option is set ``True``, ``Bilinear`` function does
-    not have linear parameters, that is, the last three term is omitted
-    and only :math:`W` works as the parameter.
-
-    .. note::
-
-       ``Bilinear`` function accepts an input variable of a non-matrix array.
-       In this case, the leading dimension is treated as the batch dimension,
-       and the other dimensions are reduced to one dimension.
-
-    .. note::
-
-       In the original paper, :math:`J` and :math:`K`
-       must be equal and the author denotes :math:`[V^1 V^2]`
-       (concatenation of matrices) by :math:`V`.
-
-    Args:
-        left_size (int): Dimension of input vector :math:`e^1` (:math:`J`)
-        right_size (int): Dimension of input vector :math:`e^2` (:math:`K`)
-        out_size (int): Dimension of output vector :math:`y` (:math:`L`)
-        nobias (bool): If ``True``, linear parameters are omitted.
-        initialW (3-D Array): Initial value of :math:`W`.
-            Shape of this argument must be
-            ``(left_size, right_size, out_size)``. If ``None``,
-            :math:`W` is initialized by centered Gaussian distribution properly
-            scaled according to the dimension of inputs and outputs.
-        initial_bias (tuple): Intial values of :math:`V^1`, :math:`V^2`
-            and :math:`b`. The length this argument must be 3.
-            Each element of this tuple must have the shapes of
-            ``(left_size, output_size)``, ``(right_size, output_size)``,
-            and ``(output_size,)``, respectively. If ``None``, :math:`V^1`
-            and :math:`V^2` is initialized by scaled centered Gaussian
-            distributions and :math:`b` is set to :math:`0`.
-
-    See:
-        `Reasoning With Neural Tensor Networks for Knowledge Base Completion
-        <http://papers.nips.cc/paper/5028-reasoning-with-neural-tensor-
-        networks-for-knowledge-base-completion>`_ [Socher+, NIPS2013].
-
-    """
-    def __init__(self, left_size, right_size, out_size, nobias=False,
-                 initialW=None, initial_bias=None):
-        super(Bilinear, self).__init__()
-        self.in_sizes = (left_size, right_size)
-        self.nobias = nobias
-
-        if initialW is not None:
-            assert initialW.shape == (left_size, right_size, out_size)
-        else:
-            # TODO(Kenta OONO): I do not know appropriate way of
-            # initializing weights in tensor network.
-            # This initialization is a modification of
-            # that of Linear function.
-            in_size = left_size * right_size * out_size
-            initialW = numpy.random.normal(
-                0, math.sqrt(1. / in_size), (left_size, right_size, out_size)
-            ).astype(numpy.float32)
-        self.params['W'] = variable.Variable(initialW)
-
-        if not self.nobias:
-            if initial_bias is not None:
-                V1, V2, b = initial_bias
-                assert V1.shape == (left_size, out_size)
-                assert V2.shape == (right_size, out_size)
-                assert b.shape == (out_size,)
-            else:
-                V1 = numpy.random.normal(
-                    0, math.sqrt(1. / left_size), (left_size, out_size)
-                ).astype(numpy.float32)
-                V2 = numpy.random.normal(
-                    0, math.sqrt(1. / right_size), (right_size, out_size)
-                ).astype(numpy.float32)
-                b = numpy.zeros(out_size, dtype=numpy.float32)
-            self.params['V1'] = variable.Variable(V1)
-            self.params['V2'] = variable.Variable(V2)
-            self.params['b'] = variable.Variable(b)
-
-    def __call__(self, e1, e2):
-        if self.nobias:
-            return bilinear(e1, e2, self.params['W'])
-        else:
-            return bilinear(e1, e2, self.params['W'], self.params['V1'],
-                            self.params['V2'], self.params['b'])
