+++ chainer/functions/connection/linear.py
@@ -63,8 +63,8 @@ class LinearFunction(function.Function):
 def linear(x, W, b=None):
     """Linear function, a.k.a. fully-connected layer.
 
-    It accepts two or three arguments: an input minibatch ``x``, an weight
-    matrix ``W``, and optionally a bias vector ``b``, and computes
+    It accepts two or three arguments: an input minibatch ``x``, a weight
+    matrix ``W``, and optionally a bias vector ``b``. It computes
     :math:`Y = xW^\top + b`.
 
     Args:
@@ -84,47 +84,3 @@ def linear(x, W, b=None):
         return LinearFunction()(x, W)
     else:
         return LinearFunction()(x, W, b)
-
-
-class Linear(link.Link):
-
-    """Linear function with attached parameters.
-
-    This model holds a weight matrix ``W`` and optionally a bias vector ``b``.
-    The weight matrix ``W`` has shape ``(out_size, in_size)``. This matrix is
-    initialized with i.i.d. Gaussian samples, each of which has zero mean and
-    deviation :math:`\sqrt{1/\\text{in_size}}`. The bias vector ``b`` is of
-    size ``out_size``. Each element is initialized with the ``bias`` value.
-    If ``nobias`` argument is set to True, then this model does not hold a bias
-    vector.
-
-    Args:
-        in_size (int): Dimension of input vectors.
-        out_size (int): Dimension of output vectors.
-        wscale (float): Scaling factor of the weight matrix.
-        bias (float): Initial bias value.
-        nobias (bool): If True, then this function does not use the bias.
-        initialW (2-D array): Initial weight value. If ``None``, then this
-            function uses to initialize ``wscale``.
-        initial_bias (1-D array): Initial bias value. If ``None``, then this
-            function uses to initialize ``bias``.
-
-    .. seealso:: :func:`~chainer.functions.linear`
-
-    """
-    def __init__(self, in_size, out_size, wscale=1, bias=0, nobias=False,
-                 initialW=None, initial_bias=None):
-        super(Linear, self).__init__()
-        if initialW is None:
-            initialW = numpy.random.normal(
-                0, wscale * math.sqrt(1. / in_size),
-                (out_size, in_size)).astype(numpy.float32)
-        self.params['W'] = variable.Variable(initialW)
-
-        if not nobias:
-            if initial_bias is None:
-                initial_bias = numpy.full(out_size, bias, dtype=numpy.float32)
-            self.params['b'] = variable.Variable(initial_bias)
-
-    def __call__(self, x):
-        return linear(x, self.params['W'], self.params.get('b', None))
