+++ tests/chainer_tests/links_tests/connection_tests/test_embed_id.py
@@ -5,8 +5,8 @@ import six
 
 import chainer
 from chainer import cuda
-from chainer import functions
 from chainer import gradient_check
+from chainer import links
 from chainer import testing
 from chainer.testing import attr
 from chainer.testing import condition
@@ -15,7 +15,7 @@ from chainer.testing import condition
 class TestEmbedID(unittest.TestCase):
 
     def setUp(self):
-        self.func = functions.EmbedID(3, 2)
+        self.func = links.EmbedID(3, 2)
         self.func.zerograds()
 
         self.W = self.func.params['W'].data.copy()  # fixed on CPU
