+++ tests/chainer_tests/links_tests/connection_tests/test_parameter.py
@@ -3,7 +3,7 @@ import unittest
 import numpy
 
 from chainer import cuda
-from chainer import functions
+from chainer import links
 from chainer import testing
 from chainer.testing import attr
 
@@ -14,7 +14,7 @@ class TestParameter(unittest.TestCase):
         self.W = numpy.random.uniform(-1, 1, (4, 3)).astype(numpy.float32)
         self.gW = numpy.random.uniform(-1, 1,
                                        self.W.shape).astype(numpy.float32)
-        self.func = functions.Parameter(self.W)
+        self.func = links.Parameter(self.W)
 
     def tearDown(self):
         del self.func
@@ -78,7 +78,7 @@ class TestVolatile(unittest.TestCase):
 
     def setUp(self):
         self.W = numpy.random.uniform(-1, 1, (4, 3)).astype(numpy.float32)
-        self.func = functions.Parameter(self.W)
+        self.func = links.Parameter(self.W)
 
     def tearDown(self):
         del self.func
