+++ tests/chainer_tests/links_tests/loss_tests/test_hierarchical_softmax.py
@@ -5,8 +5,8 @@ import numpy
 
 import chainer
 from chainer import cuda
-from chainer import functions
 from chainer import gradient_check
+from chainer import links
 from chainer import testing
 from chainer.testing import attr
 from chainer.testing import condition
@@ -16,16 +16,16 @@ class TestHuffmanTree(unittest.TestCase):
 
     def test_empty(self):
         with self.assertRaises(ValueError):
-            functions.BinaryHierarchicalSoftmax.create_huffman_tree({})
+            links.BinaryHierarchicalSoftmax.create_huffman_tree({})
 
     def test_simple(self):
-        tree = functions.BinaryHierarchicalSoftmax.create_huffman_tree(
+        tree = links.BinaryHierarchicalSoftmax.create_huffman_tree(
             {'x': 8, 'y': 6, 'z': 5, 'w': 4, 'v': 3})
         expect = (('z', 'y'), (('v', 'w'), 'x'))
         self.assertEqual(expect, tree)
 
     def test_same_count(self):
-        tree = functions.BinaryHierarchicalSoftmax.create_huffman_tree(
+        tree = links.BinaryHierarchicalSoftmax.create_huffman_tree(
             {'x': 1, 'y': 2, 'z': 3})
         # Order of the same items are not defined.
         self.assertTrue((('x', 'y'), 'z') == tree or
@@ -36,7 +36,7 @@ class TestBinaryHierarchicalSoftmax(unittest.TestCase):
 
     def setUp(self):
         tree = ((0, 1), ((2, 3), 4))
-        self.func = functions.BinaryHierarchicalSoftmax(3, tree)
+        self.func = links.BinaryHierarchicalSoftmax(3, tree)
         self.func.zerograds()
         self.x = numpy.random.uniform(-1, 1, (2, 3)).astype(numpy.float32)
         self.t = numpy.array([0, 2]).astype(numpy.int32)
